Implement getSquareBoundaryIntersection()
The math here was produced by drawing a square and a triangle on a sheet of
paper. The sign correction was produced by trial and error.

Just like getCircleBoundaryIntersection(), this function first calculate the
angle which we are looking at. This angle is then used to calculate the distance
of the square's boundary to its center. Based on this, the x and y delta to the
square's outline is computed.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>