New feature: Save/Load layout of documents to/from file.
Before rendering the APT PN/LTS, a new extension called "properties" is
added to each model element. It's value is a string representation of
the model element's associated GraphicalElement's attributes.

When parsing the APT PN/LTS apt already parses the "properties"
extension. The contained string is then parsed and all applicable
attributes are set on the GraphicalElement associated with each model
element that has such a "properties" extension.

Note: The feature is currently disabled as the user is supposed to have
the choice to NOT save layout information. Until I find a suitable way
to do so in the GUI the crucial method call in AptDocumentRenderer is
commented out.