Implement a Graphviz/Dot layout
The (already existing) random layout assigns random positions to elements. This
commit adds a new layout which uses the dot program from the Grapvhiz tool set
to compute a layout.

This commit just adds the minimum necessary to get dot working. There are a
couple of obvious improvements that might be wanted:

- Support other graphviz tools, e.g. neato, twopi, circo, fdp, sfdp
- Honor EditingOptions (Snap-to-Grid; How?)
- Configurable path to the tool

To explain the last item a bit: Right now this uses a ProcesBuilder to start
"dot". This works on Linux if "dot" is in $PATH. On Windows this will break for
sure. To fix this, the path to the "dot" tool would need to be configurable.
Somehow...

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>